feat: move queryKey logic into hook , + force using WF_BASE_QUERYKEY
