+++ hat/assets/js/apps/Iaso/domains/validationWorkflowsConfiguration/api/Delete.ts
@@ -1,4 +1,9 @@
+import React from 'react';
 import { UseMutationResult } from 'react-query';
+import {
+    getApiValidationWorkflowsListQueryKey,
+    getApiValidationWorkflowsRetrieveQueryKey,
+} from 'Iaso/api/validationWorkflows';
 import { deleteRequest } from 'Iaso/libs/Api';
 import { useSnackMutation } from 'Iaso/libs/apiHooks';
 import { API_URL, WF_BASE_QUERYKEY } from '../constants';
@@ -12,10 +17,19 @@ const deleteNode = ({
 }) => deleteRequest(`${API_URL}${workflowSlug}/node-templates/${nodeSlug}/`);
 
 export const useDeleteNode = (
-    invalidateQueryKey?: string[],
+    workflowSlug?: string,
 ): UseMutationResult<any, any> => {
+    const queryKey = React.useMemo(() => {
+        return [
+            ...getApiValidationWorkflowsListQueryKey(),
+            ...(workflowSlug
+                ? getApiValidationWorkflowsRetrieveQueryKey(workflowSlug)
+                : []),
+        ] as string[];
+    }, [workflowSlug]);
+
     return useSnackMutation({
         mutationFn: deleteNode,
-        invalidateQueryKey: invalidateQueryKey ?? [WF_BASE_QUERYKEY],
+        invalidateQueryKey: [...queryKey, WF_BASE_QUERYKEY],
     });
 };
