+++ hat/assets/js/apps/Iaso/domains/validationWorkflowsConfiguration/components/CreateEditNode/CreateEditNode.tsx
@@ -7,10 +7,6 @@ import {
     useSafeIntl,
 } from 'bluesquare-components';
 import { Field, FormikProvider, useFormik } from 'formik';
-import {
-    getApiValidationWorkflowsListQueryKey,
-    getApiValidationWorkflowsRetrieveQueryKey,
-} from 'Iaso/api/validationWorkflows';
 import { EditIconButton } from 'Iaso/components/Buttons/EditIconButton';
 import { BooleanInput } from 'Iaso/components/forms/BooleanInput';
 import { MultiSelect } from 'Iaso/domains/pages/components/MultiSelect';
@@ -49,16 +45,7 @@ export const CreateEditNode: FunctionComponent<Props> = ({
     const { data: userRoles, isFetching: isLoadingRoles } =
         useGetUserRolesDropDown();
 
-    const queryKey = React.useMemo(() => {
-        return [
-            ...getApiValidationWorkflowsListQueryKey(),
-            ...(node?.slug
-                ? getApiValidationWorkflowsRetrieveQueryKey(node?.slug)
-                : []),
-        ] as string[];
-    }, [node?.slug]);
-
-    const { mutateAsync: save } = useSaveNode(queryKey);
+    const { mutateAsync: save } = useSaveNode(node?.slug);
     const validationSchema = useNodeValidation();
     const formik = useFormik<NodeFormValues>({
         initialValues: {
